+++ lib/testcentricity_web/data_objects/excel_helper.rb
@@ -1,6 +1,3 @@
-require 'time'
-require 'chronic'
-require 'faker'
 require 'spreadsheet'
 
 
@@ -273,27 +270,6 @@ module TestCentricity
       # rename new Excel document, replacing the original
       File.rename(outfile, file)
     end
-
-    private
-
-    def self.calculate_dynamic_value(value)
-      test_value = value.split('!', 2)
-      parameter = test_value[1].split('.', 2)
-      case parameter[0]
-      when 'Date'
-        result = eval("Chronic.parse('#{parameter[1]}')")
-      when 'FormattedDate', 'FormatDate'
-        date_time_params = parameter[1].split(' format! ', 2)
-        date_time = eval("Chronic.parse('#{date_time_params[0].strip}')")
-        result = date_time.to_s.format_date_time("#{date_time_params[1].strip}")
-      else
-        result = if Faker.constants.include?(parameter[0].to_sym)
-                   eval("Faker::#{parameter[0]}.#{parameter[1]}")
-                 else
-                   eval(test_value[1])
-                 end
-      end
-      result.to_s
-    end
   end
 end
+